backend/requirements.txt
addcomments==0.0.3
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.5.30
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
daphne==3.0.2
Django==3.2.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-comment-migrate==0.1.5
django-cors-headers==3.10.1
django-filter==21.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-restql==0.15.1
django-simple-captcha==0.5.14
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
dnspython==2.3.0
drf-yasg==1.20.0
dvadmin-celery==1.0.3
et-xmlfile==1.1.0
eventlet==0.33.3
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
hiredis==2.1.1
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpath==0.82
kombu==5.2.4
MarkupSafe==2.0.1
msgpack==1.0.4
mysqlclient==2.0.3
openpyxl==3.0.9
packaging==20.9
Pillow==8.3.1
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==23.0.0
pyparsing==2.4.7
pyPEG2==2.15.2
pypinyin==0.46.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
service-identity==21.1.0
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
tenant-schemas-celery==2.0.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==3.10.0.0
tzdata==2022.7
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
selenium==4.15.0
playwright==1.40.0
pytest-playwright==0.4.0
whitenoise==5.3.0
zope.event==4.6
zope.interface==5.5.2
